Shift SQS functionality and create helpers module
Why these changes are being introduced:
* SQS-related functionality should be shifted from the CLI to the Article class. Several modules can also be consolidated into a helpers module.

How this addresses that need:
* Consolidate crossref, s3, ses, sqs, and wiley modules into new helpers module along with corresponding unit tests
* Shift SQS functionality from CLI to Article class with new attributes and upload_files_and_send_sqs_message method
* Update database unit test name to reflect new method name

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-898
