--- tests/test_crossref.py
@@ -1,15 +1,0 @@
-from awd import crossref
-
-
-def test_get_dois_from_spreadsheet():
-    dois = crossref.get_dois_from_spreadsheet("tests/fixtures/doi_success.csv")
-    for doi in dois:
-        assert doi == "10.1002/term.3131"
-
-
-def test_get_crossref_work_from_doi(mocked_web):
-    response = crossref.get_response_from_doi(
-        "http://example.com/works/", "10.1002/term.3131"
-    )
-    work = response.json()
-    assert work["message"]["title"] == ["Metal nanoparticles for bone tissue engineering"]
